backend/requirements.txt
# Core ML/DL Frameworks
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Transformers and Protein Models
transformers==4.37.0
fair-esm==2.0.0
accelerate==0.26.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6

# Protein/Bio Libraries
biopython==1.83
biotite==0.38.0
prody==2.4.0
py3Dmol==2.0.3

# Data Processing
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pyyaml==6.0.1

# Caching & Database
redis==5.0.1
sqlalchemy==2.0.25

# Monitoring & Logging
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
